Fix solution to exercise 734A

+++ codeforces/734A/anton-and-danik.cpp
@@ -1,4 +1,5 @@
 #include <bits/stdc++.h>
+
 using namespace std;
 
 int main() {
@@ -11,8 +12,8 @@ int main() {
     string s;
     cin >> s;
 
-    int A = count(s.begin(), s.end(), 'A');
-    int D = n - A;
+    long long A = count(s.begin(), s.end(), 'A');
+    long long D = n - A;
 
     if (A > D) cout << "Anton\n";
     else if (A < D) cout << "Danik\n";
